\name{arrj}
\alias{arrj}
\title{Função para arranjo com repetição}
\usage{
arrj(n,r)

}
\arguments{
numeros  numeros em geral

}
\description{
Funcao para arranjo com reposição, os elementos que compõem o conjunto podem aparecer repetidos em um agrupamento, ou seja, ocorre a repetição de um mesmo elemento em um agrupamento, onde “n” e o numero de elementos do conjunto e "r" quantidade de elementos por agrupamento, o tamanho do agrupamento que deseja.
}
\examples{
valor1 <- 8 , em que 8 é o tamanho do grupo de itens, objetos, indivíduos, números, que deseja formar grupos menores
valor2 <- 2 , em que 2 é o tamanho dos diferentes grupos de itens, objetos, indivíduos, números.. que deseja-se formar.
Digamos que possuímos uma equipe medica com 8 funcionários, e deseja-se formar um plantão com 2 dos funcionários, em que um será o plantão da manhã e o 2 será o plantão da noite, sabendo que o mesmo funcinário pode estar de manhã e a noite
arrj(valor1,valor2)



}

