\name{combnr}
\alias{combnr}
\title{Função para combinação com repetição}
\usage{
combnr(n,r)

}
\arguments{
numeros  numeros em geral
string   valor associado a um valor
}
\description{
 Função para combinação com reposição, na combinação os agrupamentos são caracterizados somente pela natureza dos elementos, combinar os n elementos que possuo, em que serão combinações de tamanho n, em que será possível combinação do mesmo elemento nas n posições. Em que o vetor a, será o vetor com itens, indivíduos, objetos, números, e outros, de interesse para combinar.
}
\examples{
Digamos que possuímos uma turma de n alunos que irão formar no curso de teatro. E deseja-se formar uma equipe de alunos da turma que serão da comissão de festa, com r funções para organizar a festa, em que é possível um mesmo aluno ter mais de uma função na comissão.
valor1 <- 25 , em que indica que a turma tem 25 anos
valor2 <- 5 , em que indica que são 5 funções da comissão, sendo possível um mesmo aluno possuir 1 ate mesmo as 5 funções.
 combnr (valor1,valor2)



}

