\name{combobj}
\alias{combobj}
\title{Função para combinação simples (n = k)}
\usage{
  combobj(n)

}
\arguments{
  numeros  numeros em geral.
  string   valor associado a um valor.
  vetor    vetor de numeros.
}
\description{
  Função para combinação com o mesmo número de objetos sem reposição, ou seja, combinar os n elementos que possuo. Em que o vetor a, será o vetor com itens, indivíduos, objetos, números, e outros, de interesse para combinar.

}
\examples{
  a <- c(“vermelho”, “branco”, “azul”) , em que indica grupo de cores, que deseja-se combinar sem reposição, combinações de tamanho 3.  Em que o resultado sempre será 1, pois não importa a ordem que apareçam os 3 elementos, a combinação será a mesma.
combobj(a)




}

