\name{combsr}
\alias{combsr}
\title{Função para combinação simples (n > k)}
\usage{
  combsr(n,r)

}
\arguments{
  numeros  numeros em geral

}
\description{
  Função para combinação sem reposição, ou seja, combinar os n elementos que possuo. Em que o vetor a, será o vetor com itens, indivíduos, objetos, números, e outros, de interesse para combinar.
}
\examples{
  combsr(n,r)
  Onde n, indica o numero total de elementos, e r, indica o numero de elementos escolhidos. Um exemplo, quer se escolher 4 dentre 9 camisetas, de quantos modos distintos podemos fazer isso, teremos então nosso n=9, e nosso 5=4, e assim.
  combsr(n,r)

}

