\name{pmr}
\alias{pmr}
\title{Função para permutação com repetição}
\usage{
pmr(b,a)

}
\arguments{
numeros  numeros em geral
string   valor associado a um valor
}
\description{
Funcao para permutação com repetição usando temos um conjunto de "b" objetos, pessoas ou coisas, que possua "a" elementos repetidos, então para calcular todas as possíveis trocas dessas "b" coisas, vamos utilizar permutação com repetição.
Para o uso o valor de b deve se ser sempre maior que o valor de a.
}
\examples{
pmr(8,2)

valor1 <- 8
valor2 <- 2
pmr(valor1,valor2)




}

