\name{pn}
\alias{pn}
\title{Função para permutação simples}
\usage{
pn(valor)

}
\arguments{
numeros  numeros em geral

}
\description{
Função para permutação simples em que o "valor" será a quantidade de "n" algarismo da questão.

}
\examples{
pn(5)

pn(5)
valor <- 5 , em que indica que 5 pessoas diferentes querem tirar uma foto, de quantas maneiras distintas as 5 pessoas podem se posicionar lado a lado, sem restrições.
pn(valor)




}


